공지사항 수정 로직 변경

+++ backend/src/main/java/place/skillexchange/backend/service/FileServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import place.skillexchange.backend.dto.FileDto;
+import place.skillexchange.backend.dto.NoticeDto;
 import place.skillexchange.backend.entity.File;
 import place.skillexchange.backend.entity.Notice;
 import place.skillexchange.backend.entity.User;
@@ -55,6 +56,14 @@ public class FileServiceImpl implements FileService{
             file = user.getFile();
             if (file != null) {
                 fileRepository.delete(file);
+
+                String fileUrl = file.getFileUrl(); // File 객체에서 URL을 가져옴
+                URL url = new URL(fileUrl); // URL 객체 생성
+                String filePath = url.getPath(); // URL의 경로 부분을 추출
+                // 맨 앞의 '/' 제거
+                String cleanedPath = filePath.substring(1);
+                //s3에서 이미지 삭제
+                s3Uploader.delete(cleanedPath);
             }
         }
 
@@ -85,26 +94,44 @@ public class FileServiceImpl implements FileService{
      * 다중 파일 업로드 ( 공지사항 수정 )
      */
     @Override
-    public List<File> updateNoticeImg(List<MultipartFile> multipartFiles, Notice notice) throws IOException {
+    public List<File> updateNoticeImg(List<String> imgUrl, List<MultipartFile> multipartFiles, Notice notice) throws IOException {
         List<File> updatedImages  = new ArrayList<>();
 
-        //이전에 저장한 이미지가 있다면 삭제
+        //이전에 저장했던 이미지 조회
         List<File> files = fileRepository.findAllByNotice(notice);
+        //저장했던 이미지가 있다면
         if (files != null && !files.isEmpty()) {
+            //저장했던 이미지들 for문
             for (File file : files) {
-                //db에서 이미지 삭제
-                fileRepository.delete(file);
-
-                String fileUrl = file.getFileUrl(); // File 객체에서 URL을 가져옴
-                URL url = new URL(fileUrl); // URL 객체 생성
-                String filePath = url.getPath(); // URL의 경로 부분을 추출
-                // 맨 앞의 '/' 제거
-                String cleanedPath = filePath.substring(1);
-                //s3에서 이미지 삭제
-                s3Uploader.delete(cleanedPath);
+                //저장했던 이미지Url
+                String fileUrl = file.getFileUrl();
+                //이미지Url이 요청이 들어온 이미지Url들과 동일한지 비교하기 위한 논리 자료형
+                boolean isUrlFoundInImgUrl = false;
+                //요청이 들어온 이미지Url들 for문
+                for (String imgUrlItem : imgUrl) {
+                    //db에 저장했던 Url과 요청 들어온 이미지Url이 동일하다면
+                    if (fileUrl.equals(imgUrlItem)) {
+                        //반환 List에 추가(ResponseDto로 쓰일 것)
+                        updatedImages.add(file);
+                        //동일하므로 논리자료형 true
+                        isUrlFoundInImgUrl = true;
+                        //저장했던 이미지Url과 요청이 들어온 Url이 동일한 것을 확인했으므로 break, 첫번째 for문으로 돌아감
+                        break;
+                    }
+                }
+                // 동일하지 않은 이미지 삭제 작업
+                if (!isUrlFoundInImgUrl) {
+                    // db에서 이미지 삭제
+                    fileRepository.delete(file);
+                    // S3에서 이미지 삭제
+                    URL url = new URL(fileUrl);
+                    String filePath = url.getPath();
+                    String cleanedPath = filePath.substring(1);
+                    s3Uploader.delete(cleanedPath);
+                }
             }
-        }
 
+        }
 
         // 수정 시 이미지 첨부를 한 경우
         if (multipartFiles != null && !multipartFiles.isEmpty()) {
@@ -117,9 +144,9 @@ public class FileServiceImpl implements FileService{
                     updatedImages.add(newFile);
                 }
             }
+            //modDate() : 수정일 현재 시간으로 업데이트
             notice.updateModDate();
         }
-
         return updatedImages;
     }
